# backend/internal/httpserver/openapi.yaml
openapi: 3.0.3
info:
  title: Timemachine Auto API
  version: "1.0.0"
  description: |
    API-first backend for the «Машина времени» auto service: public catalog,
    online booking, MAX messenger integration and admin mini-CRM for prices.
servers:
  - url: /api/v1
    description: relative to the backend origin
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /admin/prices/jobs/{id}/apply:
    post: { summary: Apply a validated import job, security: [ { bearerAuth: [] } ], responses: { '200': { description: applied } } }
...
security:
  - bearerAuth: []